Add authorization checks for updating and deleting contacts

+++ controllers/contactController.js
@@ -50,6 +50,13 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error(
+      "User don't have the permission to update other user contacts"
+    );
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -67,7 +74,15 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
-  await Contact.deleteOne({ _id: req.params.id }); // Contact.remove()
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error(
+      "User don't have the permission to delete other user contacts"
+    );
+  }
+
+  await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
